+++ packages/core-js-compat/src/modules-by-versions.mjs
@@ -112,6 +112,7 @@ export default {
     'esnext.typed-array.from-async',
   ],
   '3.20': [
+    'esnext.function.un-this',
     'esnext.iterator.to-async',
   ],
 };