+++ packages/core-js/features/function/un-this.js
@@ -1,0 +1,4 @@
+require('../../modules/esnext.function.un-this');
+var entryUnbind = require('../../internals/entry-unbind');
+
+module.exports = entryUnbind('Function', 'unThis');