+++ packages/core-js/features/function/virtual/index.js
@@ -1,3 +1,4 @@
 var parent = require('../../../stable/function/virtual');
+require('../../../modules/esnext.function.un-this');
 
 module.exports = parent;