+++ packages/core-js/proposals/function-un-this.js
@@ -1,0 +1,2 @@
+// https://github.com/js-choi/proposal-function-un-this
+require('../modules/esnext.function.un-this');